% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusMaps.R
\name{censusMapsServer}
\alias{censusMapsServer}
\alias{censusMapsInput}
\alias{censusMapsOutput}
\alias{censusMapsUI}
\alias{censusMapsApp}
\title{Shiny Server for censusMaps Package}
\usage{
censusMapsServer(id, census_geometry)

censusMapsInput(id)

censusMapsOutput(id)

censusMapsUI(id)

censusMapsApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{census_geometry}{static data frames}
}
\value{
reactive server

nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Server for censusMaps Package

Shiny Module Input for censusMaps

Shiny Module Output for censusMaps

Shiny Module UI for censusMaps

Shiny Module App for censusMaps
}
