% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cession.R
\name{cession_intersection}
\alias{cession_intersection}
\alias{cession_expand}
\alias{cession_places}
\alias{ggplot_cession}
\alias{tmap_cession}
\title{Title}
\usage{
cession_intersection(object, treaty)

cession_expand(cessions, nativeLand)

cession_places(cessions = NULL, treaty = NULL)

ggplot_cession(cessions, treaty = NULL, title = "")

tmap_cession(cessions, treaty = NULL)
}
\arguments{
\item{object}{simple feature}

\item{treaty}{simple feature}

\item{cessions}{simple feature}

\item{nativeLand}{simple feature}

\item{title}{title of plot}
}
\value{
simple feature

simple feature

simple feature

ggplot object

thematic map
}
\description{
Title

Expand Cessions beyond Treaty boundary

Title

Title

Title
}
