% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_reform.R
\name{features_reform}
\alias{features_reform}
\title{Reform Features as valid SF object}
\usage{
features_reform(features)
}
\arguments{
\item{features}{object from native-land.ca}
}
\value{
valid SF object
}
\description{
Reform Features as valid SF object
}
