% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nativeLand.R
\name{get_nativeLand}
\alias{get_nativeLand}
\title{Get Native Land information from native-land.ca}
\usage{
get_nativeLand(
  categories = c("territories", "languages", "treaties"),
  name = NULL,
  key = NULL,
  slug = NULL
)
}
\arguments{
\item{categories}{one or more categories}

\item{name}{name to look up (exact unless `slug` provided)}

\item{key}{API key for `native-land.ca`}

\item{slug}{lookup table for slug name}
}
\value{
simple feature using `sf` package
}
\description{
Get Native Land information from native-land.ca
}
