% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggplotServer}
\alias{ggplotServer}
\alias{ggplotInput}
\alias{ggplotOutput}
\alias{ggplotApp}
\title{Shiny Server for ggplot Example}
\usage{
ggplotServer(id, places = shiny::reactive(NULL))

ggplotInput(id)

ggplotOutput(id)

ggplotApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{places}{reactive object with places}

\item{input, output, session}{shiny server reactives}
}
\value{
reactive server

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Server for ggplot Example

Shiny Module Input for ggplot

Shiny Module Output for ggplot

Shiny Module App for ggplot
}
