% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_nativeLand.R
\name{ggplot_nativeLand}
\alias{ggplot_nativeLand}
\title{ggplot of nativeLand data}
\usage{
ggplot_nativeLand(object, color = NULL, title = title_names, label = TRUE)
}
\arguments{
\item{object}{simple features object}

\item{color}{vector of color}

\item{title}{character string}

\item{label}{include labels if `TRUE`}
}
\value{
gg object
}
\description{
ggplot of nativeLand data
}
