% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_sf.R
\name{ggplot_sf}
\alias{ggplot_sf}
\alias{ggplot_layer_sf}
\alias{ggplot_layer_name}
\title{Simple feature ggplot wrapper}
\usage{
ggplot_sf(legend.position = "none", color = c("grey", "black", "red"))

ggplot_layer_sf(
  object,
  color = "black",
  fill = "transparent",
  linewidth = 0.5,
  ...
)

ggplot_layer_name(object, label = "Name", color = "black", ...)
}
\arguments{
\item{legend.position}{position of legend (default "bottom")}

\item{color}{edge color}

\item{fill}{fill color}

\item{linewidth}{edge line width}

\item{...}{additional parameters}

\item{label}{column with label}
}
\value{
gg plot object

gg plot object

gg plot object
}
\description{
Simple feature ggplot wrapper

Simple feature ggplot geom_sf layer

Simple feature ggplot geom_text_reple label layer
}
