% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landGgplot.R
\name{landGgplotServer}
\alias{landGgplotServer}
\alias{landGgplotInput}
\alias{landGgplotOutput}
\alias{landGgplotApp}
\title{Shiny Server for landGgplot Example}
\usage{
landGgplotServer(id, mainpar, places = shiny::reactive(NULL))

landGgplotInput(id)

landGgplotOutput(id)

landGgplotApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{mainpar}{reactiveValue with height}

\item{places}{reactive object with places}

\item{input, output, session}{shiny server reactives}
}
\value{
reactive server

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Server for landGgplot Example

Shiny Module Input for landGgplot

Shiny Module Output for landGgplot

Shiny Module App for landGgplot
}
