% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmapr.R
\name{landmaprServer}
\alias{landmaprServer}
\alias{landmaprInput}
\alias{landmaprOutput}
\alias{landmaprUI}
\alias{landmaprApp}
\title{Shiny Server for landmapr Package}
\usage{
landmaprServer(id, nativeLandSlug, nativeLandUS, census_geometry)

landmaprInput(id)

landmaprOutput(id)

landmaprUI(id)

landmaprApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{nativeLandSlug, nativeLandUS, census_geometry}{static data frames}
}
\value{
reactive server

nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Server for landmapr Package

Shiny Module Input for landmapr

Shiny Module Output for landmapr

Shiny Module UI for landmapr

Shiny Module App for landmapr
}
