% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landMaps.R
\name{landMapsServer}
\alias{landMapsServer}
\alias{landMapsInput}
\alias{landMapsOutput}
\alias{landMapsApp}
\title{Shiny Server for landMaps Package}
\usage{
landMapsServer(id)

landMapsInput(id)

landMapsOutput(id)

landMapsApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{input, output, session}{shiny server reactives}
}
\value{
reactive server

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Server for landMaps Package

Shiny Module Input for landMaps

Shiny Module Output for landMaps

Shiny Module App for landMaps
}
