% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landPlot.R
\name{landPlotServer}
\alias{landPlotServer}
\alias{landPlotOutput}
\alias{landPlotApp}
\title{Shiny Server for landPlot Example}
\usage{
landPlotServer(id, gg_object = shiny::reactive(NULL))

landPlotOutput(id)

landPlotApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{gg_object}{reactive gg object}

\item{input, output, session}{shiny server reactives}
}
\value{
reactive server

nothing returned

nothing returned
}
\description{
Shiny Server for landPlot Example

Shiny Module Output for landPlot

Shiny Module App for landPlot
}
