% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landTable.R
\name{landTableServer}
\alias{landTableServer}
\alias{landTableOutput}
\alias{landTableApp}
\title{Shiny Server for landTable Display}
\usage{
landTableServer(id, places = shiny::reactive(NULL))

landTableOutput(id)

landTableApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{places}{reactive object with places}

\item{input, output, session}{shiny server reactives}
}
\value{
reactive server

nothing returned

nothing returned
}
\description{
Shiny Server for landTable Display

Shiny Module Output for landTable

Shiny Module App for landTable
}
