% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landTmap.R
\name{landTmapServer}
\alias{landTmapServer}
\alias{landTmapOutput}
\alias{landTmapApp}
\title{Shiny Server for Thematic Map Plot}
\usage{
landTmapServer(id, mainpar, places = shiny::reactive(NULL))

landTmapOutput(id)

landTmapApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{mainpar}{reactiveValue with height}

\item{places}{reactive object with places}

\item{input, output, session}{shiny server reactives}
}
\value{
reactive server

nothing returned

nothing returned
}
\description{
Shiny Server for Thematic Map Plot

Shiny Module Output for landTmap

Shiny Module App for landTmap
}
