% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nativeLand.R
\name{nativeLandServer}
\alias{nativeLandServer}
\alias{nativeLandInput}
\alias{nativeLandOutput}
\alias{nativeLandApp}
\title{Shiny Server for nativeLand Example}
\usage{
nativeLandServer(id, nativeLandSlug, nativeLandUS, census_geometry)

nativeLandInput(id)

nativeLandOutput(id)

nativeLandApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{nativeLandSlug, nativeLandUS, census_geometry}{static data frames}
}
\value{
reactive server

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Server for nativeLand Example

Shiny Module Input for nativeLand

Shiny Module Output for nativeLand

Shiny Module App for nativeLand
}
