% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nativeLand_states.R
\name{nativeLand_states}
\alias{nativeLand_states}
\title{Native Land that overlap Selected States}
\usage{
nativeLand_states(
  nativeLand_sf,
  census_sf,
  states = NULL,
  categories = "territories"
)
}
\arguments{
\item{nativeLand_sf}{simple feature of nativeLand data}

\item{census_sf}{simple feature of census data}

\item{states}{selected state abbreviations}

\item{categories}{one or more categories}
}
\value{
simple feature
}
\description{
Native Land that overlap Selected States
}
