% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states_nativeLand.R
\name{states_nativeLand}
\alias{states_nativeLand}
\title{States that overlap Selected Native Land}
\usage{
states_nativeLand(census_sf, nativeLand_sf, catname = NULL)
}
\arguments{
\item{census_sf}{simple feature of census data}

\item{nativeLand_sf}{simple feature of nativeLand data}

\item{catname}{selected `catname` = `category, Slug`}
}
\value{
simple feature
}
\description{
States that overlap Selected Native Land
}
