% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap.R
\name{tmapServer}
\alias{tmapServer}
\alias{tmapInput}
\alias{tmapOutput}
\alias{tmapApp}
\title{Shiny Server for tmap Plot}
\usage{
tmapServer(id, places = shiny::reactive(NULL))

tmapInput(id)

tmapOutput(id)

tmapApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{places}{reactive object with places}

\item{input, output, session}{shiny server reactives}
}
\value{
reactive server

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Server for tmap Plot

Shiny Module Input for tmap

Shiny Module Output for tmap

Shiny Module App for tmap
}
