% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{censusServer}
\alias{censusServer}
\alias{censusInput}
\alias{censusOutput}
\alias{censusApp}
\title{Shiny Server for census Example}
\usage{
censusServer(id)

censusInput(id)

censusOutput(id)

censusApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{input, output, session}{shiny server reactives}
}
\value{
reactive server

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Server for census Example

Shiny Module Input for census

Shiny Module Output for census

Shiny Module App for census
}
