% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_nativeLand.R
\name{ggplot_nativeLand}
\alias{ggplot_nativeLand}
\title{ggplot of nativeLand data}
\usage{
ggplot_nativeLand(object, colors = NULL, title = title_names)
}
\arguments{
\item{object}{simple features object}

\item{colors}{vector of colors}

\item{title}{character string}
}
\value{
gg object
}
\description{
ggplot of nativeLand data
}
