% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landMaps.R
\name{landMapsServer}
\alias{landMapsServer}
\alias{landMapsInput}
\alias{landMapsOutput}
\alias{landMapsUI}
\alias{landMapsApp}
\title{Shiny Server for landMaps Package}
\usage{
landMapsServer(id, nativeLandSlug, nativeLandUS, census_geometry)

landMapsInput(id)

landMapsOutput(id)

landMapsUI(id)

landMapsApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{nativeLandSlug, nativeLandUS, census_geometry}{static data frames}
}
\value{
reactive server

nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Server for landMaps Package

Shiny Module Input for landMaps

Shiny Module Output for landMaps

Shiny Module UI for landMaps

Shiny Module App for landMaps
}
