% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landPlot.R
\name{landPlotServer}
\alias{landPlotServer}
\alias{landPlotInput}
\alias{landPlotOutput}
\alias{landPlotApp}
\title{Shiny Server for landPlot Example}
\usage{
landPlotServer(
  id,
  gg_object = shiny::reactive(NULL),
  color = shiny::reactive(NULL)
)

landPlotInput(id)

landPlotOutput(id)

landPlotApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{gg_object, color}{reactive objects}

\item{input, output, session}{shiny server reactives}
}
\value{
reactive server

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Server for landPlot Example

Shiny Module Input for landPlot

Shiny Module Output for landPlot

Shiny Module App for landPlot
}
