% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_census.R
\name{read_census}
\alias{read_census}
\title{Read census geometry}
\usage{
read_census(geography)
}
\arguments{
\item{geography}{character string}
}
\value{
simple feature object
}
\description{
Note that `geography` and `NAME` will be renamed in saved objects
as `category` and `Name`, respectively, to agree with Native Land names.
}
